// src/app/core/interfaces/signUpDataTemplate.ts
export interface SignUpDataTemplate {
  uuid: string;
  isSetupCompleted_FLAG: boolean;
  isVerified: boolean;
  fName: string;
  isPro: boolean;
  image: string;
  oneLiner: string;
  lName: string;
  email: string;
  password: string;
  description: string;
  country: string;
  subDomainPrefix: string;
}